Disable user registration
Disable user registration for this homeserver for now. New users can
register an account with matrix.org, in a probably more convenient way,
with SSO from a plethora of providers.

AFAIU, only the homeserver users can create new rooms, so we also limit
that potential way of abusing the server.

In the future, it might make sense to allow user registration for
internal members though, fe. limiting to those with a *.gov.it,
*.governo.it, or *developers.italia.it email. Let's see how it goes.